DataFetcher/requirement.txt
pymongo==4.10.1
python-dotenv==1.0.1
yfinance==0.2.50